Null dereference crash un ApplicationCacheGroup::startLoadingEntry() https://bugs.webkit.org/show_bug.cgi?id=187012 <rdar://problem/40793716>
Reviewed by Youenn Fablet.

m_entryLoader can be null because ApplicationCacheResourceLoader::create() return null when
CachedResourceLoader::requestRawResource() fails synchronously. In such case, the completion
handler gets called with a ApplicationCacheResourceLoader::Error::CannotRequestResource error.

To address the issue, we capture the request's URL in the lambda and use it instead of trying
to get the URL from the loader's resource.

* loader/appcache/ApplicationCacheGroup.cpp:
(WebCore::ApplicationCacheGroup::startLoadingEntry):


Identifier: 202262@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233171 268f45cc-cd09-0410-ab3c-d52691b4dbfc