[LFC] Adjust static position for out-of-flow positioned boxes. https://bugs.webkit.org/show_bug.cgi?id=187000
Reviewed by Antti Koivisto.

The static position of an out-of-flow positioned box is the the position where box would go
if it was in-flow positioned. This position needs to the resolved in the containing block's coordinate system.

* layout/blockformatting/BlockFormattingContext.h:
* layout/blockformatting/BlockFormattingContextGeometry.cpp:
(WebCore::Layout::BlockFormattingContext::Geometry::inFlowReplacedWidthAndMargin):
(WebCore::Layout::BlockFormattingContext::Geometry::staticPositionForOutOfFlowPositioned):
(WebCore::Layout::BlockFormattingContext::Geometry::staticPosition):

Identifier: 202279@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233188 268f45cc-cd09-0410-ab3c-d52691b4dbfc