Handle the case of registerMDNSNameCallback called several times https://bugs.webkit.org/show_bug.cgi?id=187150 <rdar://problem/41329832>
Reviewed by Eric Carlson.

This is a speculative fix on the basis that registerMDNSNameCallback may be called several times.
In that case, we would have freed the context after the first call and would reuse it for the second call.

Instead, keep a map of pending requests and pass to registerMDNSNameCallback an identifier to that map.
If the map has no value for that identifier, return early.

* NetworkProcess/webrtc/NetworkMDNSRegister.cpp:
(WebKit::NetworkMDNSRegister::~NetworkMDNSRegister):
(WebKit::pendingRegistrationRequests):
(WebKit::registerMDNSNameCallback):
(WebKit::NetworkMDNSRegister::clearPendingRequests):
(WebKit::NetworkMDNSRegister::registerMDNSName):
* NetworkProcess/webrtc/NetworkMDNSRegister.h:
(): Deleted.


Identifier: 202415@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233328 268f45cc-cd09-0410-ab3c-d52691b4dbfc