Crash under WebCore::AXObjectCache::handleMenuItemSelected https://bugs.webkit.org/show_bug.cgi?id=186918 <rdar://problem/41365984>
Reviewed by Chris Fleizach.

Source/WebCore:

When a node is being destroyed, we deregister it from the AX cache through the Node's destructor.
But we did not remove the corresponding entry from the m_deferredFocusedNodeChange list. It would
then lead to a crash if we try to access the deleted node from m_deferredFocusedNodeChange.
Fixed it by removing the entry if the newly focused node is being destroyed.

Test: accessibility/accessibility-crash-focused-element-change.html

* accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::remove):

LayoutTests:

* accessibility/accessibility-crash-focused-element-change-expected.txt: Added.
* accessibility/accessibility-crash-focused-element-change.html: Added.


Identifier: 202475@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233390 268f45cc-cd09-0410-ab3c-d52691b4dbfc