REGRESSION (r232186): Hardware-accelerated CSS animations using steps() timing function no longer work https://bugs.webkit.org/show_bug.cgi?id=186129
Patch by Frederic Wang <fwang@igalia.com> on 2018-07-03
Reviewed by Antoine Quint.

Source/WebCore:

When the WebAnimationsCSSIntegration flag is enabled, animating the transform property with
a steps() timing function no longer works. This is because the WebAnimation code wrongly
assumes that the transform property can always be accelerated (for counterexamples, see
GraphicsLayerCA::animationCanBeAccelerated). For consistency with AnimationBase, we make
WebAnimation fallback to non-accelerated mode when RenderBoxModelObject::startAnimation
fails. This addresses the regression previously mentioned.

Test: http/wpt/css/css-animations/start-animation-001.html

* animation/KeyframeEffectReadOnly.cpp:
(WebCore::KeyframeEffectReadOnly::applyPendingAcceleratedActions): Fallback to
non-accelerated mode if startAnimation failed.

LayoutTests:

Add a test to ensure that accelerated and non-accelerated animations are properly started
when WebAnimationsCSSIntegration is enabled. In particular, consider the case of animated
transform using steps() timing function.

* http/wpt/css/css-animations/start-animation-001-expected.html: Added.
* http/wpt/css/css-animations/start-animation-001.html: Added.

Identifier: 202547@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233462 268f45cc-cd09-0410-ab3c-d52691b4dbfc